requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
gunicorn==20.0.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
